Pod::Simple::BlackBox.pm: Better printing on EBCDIC
This commit fixes the regular expression for whether to print a literal
character or an escape to work on both ASCII and EBCDIC platforms.  Now
there doesn't have to be platform-dependent code, and EBCDIC is at full
parity with ASCII.  Prior to this commit it was a pain reading the debug
output, where even the spaces are shown as "\x{40}" and not " ".

I expanded out the pattern to explicitly name every character desired.
There are shorcuts one could make, but I was afraid that some Perl
releases would have bugs in some of them.  For example [A-Z] works even
on EBCDIC platforms to match exactly the 26 uppercase English letters ,
but I don't know if it has always worked without bugs.  It seemed safest
just to list the characters.

The pattern now fills more columns than other statements in this file.
It could be made /x and split across lines if you want (but I'm not
sure if that works in all Perl releases)