lib/B/Deparse.pm: Generalize for non-ASCII platforms
This makes ASCII platform specific code generalized to non-ASCII.

I was tempted to replace the %unctrl hash with a simple string
'@ABCDEFG...' and you use the ord of the control character to look up
its representation.  But I left it as-is.  In one place we don't want
the '\c' prefix, and in another we do.