lib/B/Deparse.pm: Output WARNING_BITS in binary
This binary value was being output as just another string, which would
cause the bit patterns that coincidentally coincided with letters to be
output as those.  This is not portable to EBCDIC.  I chose to output
them as \xXX instead of octal, as I think that is the more modern thing
to do, and it's easier for me to grok the larger values when they are in
hex.